include: posix: Avoid compiler warning about return type mismatched
Cast variables to correct type to avoid compilation warnings.

+++ include/zephyr/posix/time.h
@@ -87,7 +87,7 @@ extern "C" {
 
 static inline int32_t _ts_to_ms(const struct timespec *to)
 {
-	return (to->tv_sec * MSEC_PER_SEC) + (to->tv_nsec / NSEC_PER_MSEC);
+	return (int32_t)(to->tv_sec * MSEC_PER_SEC) + (int32_t)(to->tv_nsec / NSEC_PER_MSEC);
 }
 
 int clock_gettime(clockid_t clock_id, struct timespec *ts);
